move relative coord out of API
